fix: logout user & getCurrent user

+++ src/redux/features/auth/authOperations.ts
@@ -16,8 +16,8 @@ import {
   getCurrentUserRequest,
   getCurrentUserSuccess,
 } from './authSlice';
-import apiService from '../../../services/auth-service';
-import tokenService from '../../../services/token-service';
+import apiService from '../../../services/auth/auth-service';
+import tokenService from '../../../services/auth/token-service';
 
 type Credentials = {
   name: string;
@@ -72,9 +72,11 @@ const logIn = (credentials: LoginCreds) => async (dispatch: AppDispatch) => {
 const logOut = () => async (dispatch: AppDispatch) => {
   dispatch(logoutRequest());
   try {
-    await apiService.logoutUser();
-    dispatch(logoutSuccess());
-    tokenService.removeLocalTokens();
+    const result = await apiService.logoutUser();
+    if (result.status === 204) {
+      dispatch(logoutSuccess());
+      tokenService.removeLocalTokens();
+    }
   } catch (error) {
     if (error instanceof AxiosError) {
       dispatch(logoutError(error?.message));
@@ -87,14 +89,12 @@ const getCurrent = () => async (dispatch: AppDispatch) => {
   dispatch(getCurrentUserRequest());
   try {
     const result = await apiService.getCurrentUser();
-    console.log(result);
     if (result.code === 200) {
-      dispatch(getCurrentUserSuccess(result));
+      dispatch(getCurrentUserSuccess(result.data));
     }
   } catch (error) {
     if (error instanceof AxiosError) {
       dispatch(getCurrentUserError(error?.message));
-      toast.error(error.message);
     }
   }
 };
